+++ Tests/UDDFTests/IntegrationTests/UDDFSerializationTests.swift
@@ -6,9 +6,7 @@ final class UDDFSerializationTests: XCTestCase {
     // MARK: - Date Format Round-Trip Tests
 
     func testRoundTripWithLocalDateFormat() throws {
-        var document = try UDDFBuilder()
-            .generator(name: "TestApp", version: "1.0.0")
-            .build()
+        var document = makeDocument()
         document.generator.datetime = Date()
 
         let data = try UDDFSerialization.write(document, dateFormat: .local)
@@ -19,9 +17,7 @@ final class UDDFSerializationTests: XCTestCase {
     }
 
     func testRoundTripWithUTCDateFormat() throws {
-        var document = try UDDFBuilder()
-            .generator(name: "TestApp", version: "1.0.0")
-            .build()
+        var document = makeDocument()
         document.generator.datetime = Date()
 
         let data = try UDDFSerialization.write(document, dateFormat: .utc)
@@ -29,5 +25,12 @@ final class UDDFSerializationTests: XCTestCase {
 
         XCTAssertEqual(reparsed.generator.name, document.generator.name)
         XCTAssertNotNil(reparsed.generator.datetime)
+    }
+
+    private func makeDocument() -> UDDFDocument {
+        UDDFDocument(
+            version: "3.2.1",
+            generator: Generator(name: "TestApp", version: "1.0.0")
+        )
     }
 }